checkers/cubic/requirements.txt
pwntools==4.3.1
pycryptodome==3.9.9
gornilo==0.5.0